Documentation: use mdx to check that the examples in the documentation work
There were already some examples, but the ones in the .md/.mld are
slightly different (have comments and slightly different parameters).

Create new files and set up dune rules to check that the .md/.mld
files are in sync with the examples, and update examples to test the new
files.

This requires some 'copy_files' rules because 'mdx' cannot refer to
files from its parent dir.

Signed-off-by: Edwin Török <edwin@etorok.net>